requirements.txt
Django==3.2.9
dj-database-url==0.5.0
django-cors-headers==3.10.0
django-extensions==3.1.5
django-health-check==3.16.5
django-rq==2.5.1
django-urlid-graph==0.5.9
djangorestframework==3.12.4
gunicorn==20.1.0
https://github.com/turicas/rows/archive/develop.zip?1657983463
loguru==0.5.3
psycopg2-binary==2.9.1
python-decouple==3.6
python-keycloak==0.26.1
python-keycloak==0.26.1
whitenoise==5.3.0
requests-oauthlib
numpy==1.21.0
opensearch-py==2.6.0
opensearch-dsl
pandas==2.0.3
geoip2
